First pass at moving MathTrig tests into individual test files
