+++ tests/PhpSpreadsheetTests/Calculation/Functions/MathTrig/AcothTest.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\MathTrig;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\MathTrig;
+use PHPUnit\Framework\TestCase;
+
+class AcothTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerACOTH
+     *
+     * @param mixed $expectedResult
+     * @param mixed $number
+     */
+    public function testACOTH($expectedResult, $number)
+    {
+        $result = MathTrig::ACOTH($number);
+        $this->assertEquals($expectedResult, $result, '', 1E-12);
+    }
+
+    public function providerACOTH()
+    {
+        return require 'data/Calculation/MathTrig/ACOTH.php';
+    }
+}
