+++ tests/PhpSpreadsheetTests/Calculation/MathTrigTest.php
@@ -323,136 +323,6 @@ class MathTrigTest extends TestCase
     }
 
     /**
-     * @dataProvider providerMROUND
-     *
-     * @param mixed $expectedResult
-     */
-    public function testMROUND($expectedResult, ...$args)
-    {
-        Calculation::setArrayReturnType(Calculation::RETURN_ARRAY_AS_VALUE);
-        $result = MathTrig::MROUND(...$args);
-        Calculation::setArrayReturnType(Calculation::RETURN_ARRAY_AS_ARRAY);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerMROUND()
-    {
-        return require 'data/Calculation/MathTrig/MROUND.php';
-    }
-
-    /**
-     * @dataProvider providerPRODUCT
-     *
-     * @param mixed $expectedResult
-     */
-    public function testPRODUCT($expectedResult, ...$args)
-    {
-        $result = MathTrig::PRODUCT(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerPRODUCT()
-    {
-        return require 'data/Calculation/MathTrig/PRODUCT.php';
-    }
-
-    /**
-     * @dataProvider providerQUOTIENT
-     *
-     * @param mixed $expectedResult
-     */
-    public function testQUOTIENT($expectedResult, ...$args)
-    {
-        $result = MathTrig::QUOTIENT(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerQUOTIENT()
-    {
-        return require 'data/Calculation/MathTrig/QUOTIENT.php';
-    }
-
-    /**
-     * @dataProvider providerROUNDUP
-     *
-     * @param mixed $expectedResult
-     */
-    public function testROUNDUP($expectedResult, ...$args)
-    {
-        $result = MathTrig::ROUNDUP(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerROUNDUP()
-    {
-        return require 'data/Calculation/MathTrig/ROUNDUP.php';
-    }
-
-    /**
-     * @dataProvider providerROUNDDOWN
-     *
-     * @param mixed $expectedResult
-     */
-    public function testROUNDDOWN($expectedResult, ...$args)
-    {
-        $result = MathTrig::ROUNDDOWN(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerROUNDDOWN()
-    {
-        return require 'data/Calculation/MathTrig/ROUNDDOWN.php';
-    }
-
-    /**
-     * @dataProvider providerSERIESSUM
-     *
-     * @param mixed $expectedResult
-     */
-    public function testSERIESSUM($expectedResult, ...$args)
-    {
-        $result = MathTrig::SERIESSUM(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerSERIESSUM()
-    {
-        return require 'data/Calculation/MathTrig/SERIESSUM.php';
-    }
-
-    /**
-     * @dataProvider providerSUMSQ
-     *
-     * @param mixed $expectedResult
-     */
-    public function testSUMSQ($expectedResult, ...$args)
-    {
-        $result = MathTrig::SUMSQ(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerSUMSQ()
-    {
-        return require 'data/Calculation/MathTrig/SUMSQ.php';
-    }
-
-    /**
-     * @dataProvider providerSUMPRODUCT
-     *
-     * @param mixed $expectedResult
-     */
-    public function testSUMPRODUCT($expectedResult, ...$args)
-    {
-        $result = MathTrig::SUMPRODUCT(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerSUMPRODUCT()
-    {
-        return require 'data/Calculation/MathTrig/SUMPRODUCT.php';
-    }
-
-    /**
      * @dataProvider providerSUMX2MY2
      *
      * @param mixed $expectedResult
@@ -498,54 +368,6 @@ class MathTrigTest extends TestCase
     public function providerSUMXMY2()
     {
         return require 'data/Calculation/MathTrig/SUMXMY2.php';
-    }
-
-    /**
-     * @dataProvider providerTRUNC
-     *
-     * @param mixed $expectedResult
-     */
-    public function testTRUNC($expectedResult, ...$args)
-    {
-        $result = MathTrig::TRUNC(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerTRUNC()
-    {
-        return require 'data/Calculation/MathTrig/TRUNC.php';
-    }
-
-    /**
-     * @dataProvider providerROMAN
-     *
-     * @param mixed $expectedResult
-     */
-    public function testROMAN($expectedResult, ...$args)
-    {
-        $result = MathTrig::ROMAN(...$args);
-        self::assertEquals($expectedResult, $result);
-    }
-
-    public function providerROMAN()
-    {
-        return require 'data/Calculation/MathTrig/ROMAN.php';
-    }
-
-    /**
-     * @dataProvider providerSQRTPI
-     *
-     * @param mixed $expectedResult
-     */
-    public function testSQRTPI($expectedResult, ...$args)
-    {
-        $result = MathTrig::SQRTPI(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerSQRTPI()
-    {
-        return require 'data/Calculation/MathTrig/SQRTPI.php';
     }
 
     /**
@@ -753,141 +575,5 @@ class MathTrigTest extends TestCase
     public function providerNestedSUBTOTAL()
     {
         return require 'data/Calculation/MathTrig/SUBTOTALNESTED.php';
-    }
-
-    /**
-     * @dataProvider providerSEC
-     *
-     * @param mixed $expectedResult
-     * @param mixed $angle
-     */
-    public function testSEC($expectedResult, $angle)
-    {
-        $result = MathTrig::SEC($angle);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerSEC()
-    {
-        return require 'data/Calculation/MathTrig/SEC.php';
-    }
-
-    /**
-     * @dataProvider providerSECH
-     *
-     * @param mixed $expectedResult
-     * @param mixed $angle
-     */
-    public function testSECH($expectedResult, $angle)
-    {
-        $result = MathTrig::SECH($angle);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerSECH()
-    {
-        return require 'data/Calculation/MathTrig/SECH.php';
-    }
-
-    /**
-     * @dataProvider providerCSC
-     *
-     * @param mixed $expectedResult
-     * @param mixed $angle
-     */
-    public function testCSC($expectedResult, $angle)
-    {
-        $result = MathTrig::CSC($angle);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerCSC()
-    {
-        return require 'data/Calculation/MathTrig/CSC.php';
-    }
-
-    /**
-     * @dataProvider providerCSCH
-     *
-     * @param mixed $expectedResult
-     * @param mixed $angle
-     */
-    public function testCSCH($expectedResult, $angle)
-    {
-        $result = MathTrig::CSCH($angle);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerCSCH()
-    {
-        return require 'data/Calculation/MathTrig/CSCH.php';
-    }
-
-    /**
-     * @dataProvider providerCOT
-     *
-     * @param mixed $expectedResult
-     * @param mixed $angle
-     */
-    public function testCOT($expectedResult, $angle)
-    {
-        $result = MathTrig::COT($angle);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerCOT()
-    {
-        return require 'data/Calculation/MathTrig/COT.php';
-    }
-
-    /**
-     * @dataProvider providerCOTH
-     *
-     * @param mixed $expectedResult
-     * @param mixed $angle
-     */
-    public function testCOTH($expectedResult, $angle)
-    {
-        $result = MathTrig::COTH($angle);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerCOTH()
-    {
-        return require 'data/Calculation/MathTrig/COTH.php';
-    }
-
-    /**
-     * @dataProvider providerACOT
-     *
-     * @param mixed $expectedResult
-     * @param mixed $number
-     */
-    public function testACOT($expectedResult, $number)
-    {
-        $result = MathTrig::ACOT($number);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerACOT()
-    {
-        return require 'data/Calculation/MathTrig/ACOT.php';
-    }
-
-    /**
-     * @dataProvider providerACOTH
-     *
-     * @param mixed $expectedResult
-     * @param mixed $number
-     */
-    public function testACOTH($expectedResult, $number)
-    {
-        $result = MathTrig::ACOTH($number);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerACOTH()
-    {
-        return require 'data/Calculation/MathTrig/ACOTH.php';
     }
 }
