+++ tests/data/Calculation/MathTrig/ATAN2.php
@@ -5,82 +5,66 @@
 return [
     [
         '#DIV/0!',
-        0,
-        0,
+        0, 0,
     ],
     [
         0.78539816339699997,
-        1,
-        1,
+        1, 1,
     ],
     [
         -2.3561944901919998,
-        -1,
-        -1,
+        -1, -1,
     ],
     [
         2.3561944901919998,
-        -1,
-        1,
+        -1, 1,
     ],
     [
         -0.78539816339699997,
-        1,
-        -1,
+        1, -1,
     ],
     [
         1.107148717794,
-        0.5,
-        1,
+        0.5, 1,
     ],
     [
         1.8157749899219999,
-        -0.5,
-        2,
+        -0.5, 2,
     ],
     [
         0.67474094222400005,
-        1,
-        0.80000000000000004,
+        1, 0.80000000000000004,
     ],
     [
         -0.64350110879300004,
-        0.80000000000000004,
-        -0.59999999999999998,
+        0.80000000000000004, -0.59999999999999998,
     ],
     [
         -1.460139105621,
-        1,
-        -9,
+        1, -9,
     ],
     [
         0.0,
-        0.20000000000000001,
-        0,
+        0.20000000000000001, 0,
     ],
     [
         1.107148717794,
-        0.10000000000000001,
-        0.20000000000000001,
+        0.10000000000000001, 0.20000000000000001,
     ],
     [
         1.570796326795,
-        0,
-        0.20000000000000001,
+        0, 0.20000000000000001,
     ],
     [
         '#VALUE!',
-        'A',
-        0.20000000000000001,
+        'A', 0.20000000000000001,
     ],
     [
         0.78539816339699997,
-        true,
-        1,
+        true, 1,
     ],
     [
         -1.570796326795,
-        false,
-        -2.5,
+        false, -2.5,
     ],
 ];
